ResortApp/requirements.txt
aiofiles==25.1.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
cffi==2.0.0
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
structlog==25.4.0
typing-inspection==0.4.1
typing_extensions==4.15.0
unicorn==2.1.4
uvicorn==0.35.0
pandas>=2.2.0
openpyxl==3.1.2
pytz==2024.1
requests
reportlab==4.4.10

gunicorn
cachetools
